requirements.txt
streamlit
pandas
plotly
requests
python-dotenv
scikit-learn
Pillow
numpy
